[BUILD] Enable path option for ROCM, CUDA, Vulkan, simplify optional build (#1270)
