+++ cmake/modules/LLVM.cmake
@@ -2,12 +2,7 @@
 add_definitions(-DDMLC_USE_FOPEN64=0)
 
 if(NOT USE_LLVM STREQUAL "OFF")
-  if(NOT USE_LLVM STREQUAL "ON")
-    set(LLVM_CONFIG "${USE_LLVM}")
-  else()
-    set(LLVM_CONFIG "")
-  endif()
-  find_llvm()
+  find_llvm(${USE_LLVM})
   include_directories(${LLVM_INCLUDE_DIRS})
   add_definitions(${LLVM_DEFINITIONS})
   message(STATUS "Build with LLVM " ${LLVM_PACKAGE_VERSION})
