+++ cmake/modules/OpenCL.cmake
@@ -13,7 +13,6 @@ if(USE_OPENCL)
   file(GLOB RUNTIME_OPENCL_SRCS src/runtime/opencl/*.cc)
   list(APPEND TVM_RUNTIME_LINKER_LIBS ${OpenCL_LIBRARIES})
   list(APPEND RUNTIME_SRCS ${RUNTIME_OPENCL_SRCS})
-  list(APPEND COMPILER_SRCS src/codegen/opt/build_opencl_on.cc)
 else()
   list(APPEND COMPILER_SRCS src/codegen/opt/build_opencl_off.cc)
 endif(USE_OPENCL)
