+++ cmake/modules/OpenGL.cmake
@@ -13,7 +13,6 @@ if(USE_OPENGL)
   file(GLOB RUNTIME_OPENGL_SRCS src/runtime/opengl/*.cc)
   list(APPEND TVM_RUNTIME_LINKER_LIBS ${OpenGL_LIBRARIES} glfw)
   list(APPEND RUNTIME_SRCS ${RUNTIME_OPENGL_SRCS})
-  list(APPEND COMPILER_SRCS src/codegen/opt/build_opengl_on.cc)
 else(USE_OPENGL)
   list(APPEND COMPILER_SRCS src/codegen/opt/build_opengl_off.cc)
 endif(USE_OPENGL)
