+++ src/codegen/codegen_opengl.cc
@@ -9,6 +9,7 @@
 #include <vector>
 #include <string>
 #include "./codegen_opengl.h"
+#include "./build_common.h"
 #include "../runtime/thread_storage_scope.h"
 
 namespace tvm {
@@ -267,6 +268,22 @@ void CodeGenOpenGL::VisitStmt_(const Evaluate* op) {
   this->PrintIndent();
   this->stream << GetVarID(buffer) << " = " << PrintExpr(value) << ";\n";
 }
+
+runtime::Module BuildOpenGL(Array<LoweredFunc> funcs) {
+  bool output_ssa = false;
+  CodeGenOpenGL cg;
+  cg.Init(output_ssa);
+  for (LoweredFunc f : funcs) {
+    cg.AddFunction(f);
+  }
+  auto shaders = cg.Finish();
+  return OpenGLModuleCreate(shaders, "gl", ExtractFuncInfo(funcs));
+}
+
+TVM_REGISTER_API("codegen.build_opengl")
+.set_body([](TVMArgs args, TVMRetValue* rv) {
+  *rv = BuildOpenGL(args[0]);
+});
 
 }  // namespace codegen
 }  // namespace tvm
