+++ src/codegen/opt/build_metal_off.cc
@@ -1,6 +1,20 @@
 /*!
  *  Copyright (c) 2018 by Contributors
- *  Build Metal modules off
+ *  Optional module when build metal is switched to off
  */
-#define TVM_METAL_RUNTIME 0
-#include "./build_metal.h"
+#include "../codegen_source_base.h"
+#include "../../runtime/metal/metal_module.h"
+
+namespace tvm {
+namespace runtime {
+
+Module MetalModuleCreate(std::string data,
+                         std::string fmt,
+                         std::unordered_map<std::string, FunctionInfo> fmap,
+                         std::string source) {
+  LOG(WARNING) << "Metal runtime not enabled, return a source module...";
+  return codegen::DeviceSourceModuleCreate(data, fmt, fmap, "metal");
+}
+
+}  // namespace runtime
+}  // namespace tvm
