--- src/codegen/opt/build_metal_on.cc
@@ -1,6 +1,0 @@
-/*!
- *  Copyright (c) 2018 by Contributors
- *  Build Metal modules on
- */
-#define TVM_METAL_RUNTIME 1
-#include "./build_metal.h"
