--- src/codegen/opt/build_opencl_on.cc
@@ -1,6 +1,0 @@
-/*!
- *  Copyright (c) 2018 by Contributors
- *  Build opencl modules on
- */
-#define TVM_OPENCL_RUNTIME 1
-#include "./build_opencl.h"
