--- src/codegen/opt/build_opengl_on.cc
@@ -1,6 +1,0 @@
-/*!
- *  Copyright (c) 2018 by Contributors
- *  Build OpenGL modules on
- */
-#define TVM_OPENGL_RUNTIME 1
-#include "./build_opengl.h"
